% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_theta.R
\name{add_beta}
\alias{add_beta}
\title{Include beta to the parameter vector theta}
\usage{
add_beta(theta, S0)
}
\arguments{
\item{theta}{parameters (without beta)}

\item{S0}{initial population size}
}
\value{
theta including beta
}
\description{
Include beta to the parameter vector theta
}
