% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gamma}
\alias{add_gamma}
\title{Include beta to the parameter vector theta}
\usage{
add_gamma(theta)
}
\arguments{
\item{theta}{parameters (without gamma)}
}
\value{
theta including beta
}
\description{
Include beta to the parameter vector theta
}
