% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iteration.R
\name{add_iteration}
\alias{add_iteration}
\title{Add an "Iteration" column to a data frame}
\usage{
add_iteration(df, thin)
}
\arguments{
\item{df}{data frame}

\item{thin}{thinning parameter for the Markov chain}
}
\value{
data frame with a variable for iteration
}
\description{
Add an "Iteration" column to a data frame
}
