% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_MCMC.R
\name{analyze_MCMC}
\alias{analyze_MCMC}
\title{Creates figures and summary statistics of the output of the Markov chain}
\usage{
analyze_MCMC(
  MC,
  burnin = 0,
  thin,
  n_max = NULL,
  iota_dist,
  plot_id = NULL,
  path = NULL,
  save_fig = TRUE,
  do_SS = FALSE,
  theta_true,
  Y,
  coverage = TRUE
)
}
\arguments{
\item{MC}{output of the MCMC algorithm}

\item{burnin}{number of iterations to discard, the default value NULL will discard half of the draws.}

\item{thin}{thinning parameter for the Markov chain}

\item{n_max}{last draw to take into account}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}

\item{save_fig}{logical; whether to save the figures generated}

\item{do_SS}{logical; whether to analyze the summary statistics in addition to the parameters}

\item{theta_true}{true value of the parameters}

\item{Y}{observed data}

\item{coverage}{whether to determine if the credible intervals contain the true value of the parameters}
}
\value{
a list of summary statistics including posterior means, posterior quantiles, effective sample size and the acceptance rate
}
\description{
Creates figures and summary statistics of the output of the Markov chain
}
