% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_hard.R
\name{between_hard}
\alias{between_hard}
\title{Title}
\usage{
between_hard(x, low, upp)
}
\arguments{
\item{x}{vector of one or more numeric values}

\item{low}{lower bound}

\item{upp}{upper bound}
}
\value{
logical vector the same length as x
}
\description{
Title
}
