% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trajectories.R
\name{compare_trajectories}
\alias{compare_trajectories}
\title{Draw trajectories of the PDSIR and SIR}
\usage{
compare_trajectories(SIR_SI, PDSIR_SI, plot_id = NULL, path, t_end)
}
\arguments{
\item{SIR_SI}{compartment sizes of SIR process}

\item{PDSIR_SI}{compartment sizes of PD-SIR process}

\item{plot_id}{name file for the figures}

\item{path}{directory in which to save the figure}

\item{t_end}{end of observation period}
}
\value{
save figures of trajectories
}
\description{
Compare the compartment trajectories of the stochastic SIR process and of the PD-SIR process
}
