% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Fk.R
\name{compute_Fk}
\alias{compute_Fk}
\title{Compute the number of expositions per observation interval}
\usage{
compute_Fk(x, ts, S0)
}
\arguments{
\item{x}{vector of infection times}

\item{ts}{observation schedule}

\item{S0}{size of initial susceptible population}
}
\value{
number of infections per observation interval
}
\description{
Compute the number of expositions per observation interval
}
