% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM-simulate.R
\name{compute_Ik}
\alias{compute_Ik}
\title{Discrete incidence counts for infections}
\usage{
compute_Ik(x, ts)
}
\arguments{
\item{x}{vector of infection times}

\item{ts}{observation schedule}
}
\value{
a vector of the number of infections in each time interval
}
\description{
Compute the number of infections in each observation interval
}
