% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Tk.R
\name{compute_Tk_MH_fast}
\alias{compute_Tk_MH_fast}
\title{Discrete Incidence Data for Infections (Fast)}
\usage{
compute_Tk_MH_fast(x, ts)
}
\arguments{
\item{x}{vector of infection times}

\item{ts}{observation schedule}
}
\value{
a vector of the number of infections in each time interval
}
\description{
Discrete Incidence Data for Infections (Fast)
}
