% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_exp.R
\name{dexp_trunc_log}
\alias{dexp_trunc_log}
\title{Log density of the truncated exponential distribution}
\usage{
dexp_trunc_log(x, lambda, lower, upper)
}
\arguments{
\item{x}{vector of values}

\item{lambda}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
log density
}
\description{
Log density of the truncated exponential distribution
}
\seealso{
Other truncated exponential: 
\code{\link{rexp_trunc}()}
}
\concept{truncated exponential}
