% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_acf.R
\name{draw_acf}
\alias{draw_acf}
\title{Generate a plot of the auto-correlation function}
\usage{
draw_acf(df, var, plot_id, path = NULL, lag_max = 500)
}
\arguments{
\item{df}{data frame}

\item{var}{variable under consideration}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}

\item{lag_max}{largest lag shown}
}
\value{
Barplot of the ACF
}
\description{
Generate a plot of the auto-correlation function
}
