% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_density_2d.R
\name{draw_density_2d}
\alias{draw_density_2d}
\title{Generates a bi-dimensional density plot}
\usage{
draw_density_2d(
  df,
  var1,
  var2,
  plot_id,
  path = NULL,
  xlim = c(NA_real_, NA_real_)
)
}
\arguments{
\item{df}{data frame}

\item{var1}{variable to plot on the x-axis}

\item{var2}{variable to plot on the y-axis}

\item{plot_id}{name file for the figure}

\item{path}{directory in which to save the figure}

\item{xlim}{vector of length 2; limits on the x-axis}
}
\value{
a figure of a biavriate density
}
\description{
Generates a bi-dimensional density plot
}
