% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_E3.R
\name{draw_E3_facet}
\alias{draw_E3_facet}
\title{Title}
\usage{
draw_E3_facet(results, path, var, lab_y, plot_id)
}
\arguments{
\item{results}{df of results from experiment 3}

\item{path}{directory in which to save the figure}

\item{var}{variable of the df results to visualize}

\item{lab_y}{label of y-axis}

\item{plot_id}{name file for the figure(s)}
}
\value{
saves a figures
}
\description{
Title
}
