% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_I.R
\name{draw_Ebola_data}
\alias{draw_Ebola_data}
\title{Barplot of the Observed Infection Counts}
\usage{
draw_Ebola_data(path_data = "Input/Ebola/Guinea_modified.csv", path_figure)
}
\arguments{
\item{path_data}{directory in which the .csv file is located}

\item{path_figure}{directory in which to save the figure}
}
\value{
Barplot of infection counts.
}
\description{
Barplot of the Observed Infection Counts
}
