% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_gamma.R
\name{draw_gamma}
\alias{draw_gamma}
\title{Illustrates minorization of gamma densities in manuscript}
\usage{
draw_gamma(path)
}
\arguments{
\item{path}{directory in which to save the figure}
}
\value{
Illustration of gamma minorization
}
\description{
Illustrates minorization of gamma densities in manuscript
}
