% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_histogram.R
\name{draw_histogram}
\alias{draw_histogram}
\title{Generates a histogram}
\usage{
draw_histogram(df, var, plot_id, path = NULL, bins = 30)
}
\arguments{
\item{df}{data frame}

\item{var}{variable under consideration}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}

\item{bins}{number of bins in the histogram}
}
\value{
a histogram
}
\description{
Generates a histogram
}
