% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hpd_2d.R
\name{draw_hpd_2d}
\alias{draw_hpd_2d}
\title{draws highest posterior density}
\usage{
draw_hpd_2d(df, var1, var2, true1, true2, plot_id, path = NULL, levels)
}
\arguments{
\item{df}{data frame}

\item{var1}{variable to plot on the x-axis}

\item{var2}{variable to plot on the y-axis}

\item{true1}{true value of the parameter corresponding to variable 1}

\item{true2}{true value of the parameter corresponding to variable 2}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}

\item{levels}{c(0.50, 0.95); level(s) at which to draw the HPD regions}
}
\value{
Contour plot with HPD region(s) at the specified level(s)
}
\description{
need to find better alternative because the HPDregionplot function is very buggy. check Peter Hoff's (2009) book
}
