% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mu.R
\name{draw_mu}
\alias{draw_mu}
\title{Infection dynamics in the SIR and the PD-SIR processes.}
\usage{
draw_mu(
  theta = list(R0 = 2, gamma = 1),
  S0 = 10,
  I0 = 2,
  t_end = 3,
  K = 5,
  iota_dist = "exponential",
  gener = FALSE,
  b = 1,
  path
)
}
\arguments{
\item{theta}{parameters of the SIR process}

\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{t_end}{end of observation period}

\item{K}{number of observation intervals}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}

\item{path}{directory in which to save the figure}
}
\value{
Figure contrasting the infection rate in the SIR and the PD-SIR processes.
}
\description{
Infection dynamics in the SIR and the PD-SIR processes.
}
