% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_traceplot.R
\name{draw_traceplot}
\alias{draw_traceplot}
\title{Draws a traceplot}
\usage{
draw_traceplot(df, var, plot_id, path = NULL)
}
\arguments{
\item{df}{data frame}

\item{var}{variable under consideration}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}
}
\value{
a traceplot
}
\description{
Draws a traceplot
}
