% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_trajectories.R
\name{draw_trajectories}
\alias{draw_trajectories}
\title{Trajectories of the SEM}
\usage{
draw_trajectories(SEM, plot_id = NULL, path, t_end = 10, type = "SIR")
}
\arguments{
\item{SEM}{list corresponding to stochastic epidemic process}

\item{plot_id}{name file for the figure}

\item{path}{directory in which to save the figure}

\item{t_end}{end of observation period}

\item{type}{c("SIR", "SEIR"); type of process to simulate}
}
\value{
save a figure of the compartments' trajectories
}
\description{
Line plot of the compartment sizes over time.
}
