% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{dweibull2}
\alias{dweibull2}
\title{Density of weibull distribution}
\usage{
dweibull2(x, shape, rate, log = FALSE)
}
\arguments{
\item{x}{vector}

\item{shape}{shape parameter}

\item{rate}{rate parameter}

\item{log}{logical; whether to take the logarithm of the density}
}
\value{
vector of densities
}
\description{
Density of weibull distribution
}
