% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebola_guinea_raw}
\alias{ebola_guinea_raw}
\title{Weekly infection counts in each prefecture of Guinea}
\format{
## `ebola_guinea_raw`
A list of three vectors:
\describe{
  \item{I_k}{number of weekly infections}
  \item{dates}{date of the week}
  \item{ts}{endpoints of observation intervals (units: days), used in the MCMC algorithm; the intervals correspond to the weeks in `dates`}
}
}
\usage{
ebola_guinea_raw
}
\description{
These are the raw data for Guinea
}
\keyword{datasets}
