% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_1_proof_of_concept}
\alias{experiment_1_proof_of_concept}
\title{Experiment 1: Proof of concept}
\usage{
experiment_1_proof_of_concept(
  S0 = 1000,
  I0 = 10,
  theta = list(R0 = 2.5, lambda = 1, shape = 1, gamma = 1),
  t_end = 6,
  K = 10,
  N = 1000,
  thin = 1,
  rho = 1/4,
  param = "bR",
  approx = "ldp",
  iota_dist = "exponential",
  gener = FALSE,
  b = 1/2,
  theta_0_factor = 1,
  path = NULL,
  plot_id,
  save_fig = TRUE
)
}
\arguments{
\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{theta}{parameters of the SIR process}

\item{t_end}{end of observation period}

\item{K}{number of observation intervals}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning parameter for the Markov chain}

\item{rho}{proportion of the latent data updated each iteration}

\item{param}{c("bg", "bR"); parameterize the models in terms of (beta, gamma) or (beta, R0)}

\item{approx}{c("poisson", "ldp"); whether to approximate the distribution of the infection times with a poisson process or a linear death process}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}

\item{theta_0_factor}{factors by which the true value of the parameters is multiplied to initialize the Markov chain}

\item{path}{directory in which to save the figure}

\item{plot_id}{name file for the figure(s)}

\item{save_fig}{logical; whether to save the figures generated}
}
\value{
list containing the parameters, observed data, Markov chain and run time of the algorithm
}
\description{
Experiment 1: Proof of concept
}
