% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_2_PDSIR_trajectories}
\alias{experiment_2_PDSIR_trajectories}
\title{Experiment 2: Compare trajectories of SIR and PD-SIR process}
\usage{
experiment_2_PDSIR_trajectories(
  S0 = 10000,
  I0 = 10,
  theta = list(R0 = 4, gamma = 1),
  t_end = 6,
  Ks = c(3, 5, 10, 50, 1000),
  iota_dist = "exponential",
  gener = FALSE,
  b = 1/2,
  path
)
}
\arguments{
\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{theta}{parameters of the SIR process}

\item{t_end}{end of observation period}

\item{Ks}{vector of numbers of time intervals to consider for the PD-SIR process}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}

\item{path}{directory in which to save the figure}
}
\value{
figures comparing the trajectories of the PDSIR and SIR for different value K
}
\description{
Experiment 2: Compare trajectories of SIR and PD-SIR process
}
