% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_3_acceptance_vs_rho}
\alias{experiment_3_acceptance_vs_rho}
\title{Experiment 3: Evakuate impact of rho on the M-H acceptance rate}
\usage{
experiment_3_acceptance_vs_rho(
  S0s = c(100, 500, 1000, 5000),
  I0 = 10,
  R0s = c(2, 2.5, 3, 3.5),
  gamma = 1,
  t_end = 6,
  K = 20,
  rhos = c(0.01, 0.05, 0.1, 0.25, 0.5, 0.66, 1),
  N = 1000,
  thin = 1,
  path
)
}
\arguments{
\item{S0s}{vector of sizes of initial susceptible population}

\item{I0}{size initial infectious population}

\item{R0s}{vector of R0's to be used in conjunction with S0s}

\item{gamma}{removal rate}

\item{t_end}{end of observation period}

\item{K}{number of observation intervals}

\item{rhos}{vector of values for the tuning parameter rho}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning parameter for the Markov chain}

\item{path}{directory in which to save the figure}
}
\value{
saves figures and a df summaryzing the results
}
\description{
Experiment 3: Evakuate impact of rho on the M-H acceptance rate
}
