% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_3_output_analysis}
\alias{experiment_3_output_analysis}
\title{Analyze output of MCMC runs from Experiment 3}
\usage{
experiment_3_output_analysis(output_E3, path)
}
\arguments{
\item{output_E3}{output from Experiment 3}

\item{path}{directory in which to save the figure}
}
\value{
figures of trajectories and output of Experiment 3
}
\description{
Analyze output of MCMC runs from Experiment 3
}
