% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_4_coverage}
\alias{experiment_4_coverage}
\title{Experiment 4: estimate the coverage rate of the DA-MCMC algorithm}
\usage{
experiment_4_coverage(
  S0 = 1000,
  I0 = 10,
  theta = list(R0 = 2.5, gamma = 1),
  t_end = 6,
  K = 10,
  N = 1e+05,
  thin = 1,
  rho = 1/10,
  iota_dist = "exponential",
  m = 100,
  a
)
}
\arguments{
\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{theta}{parameters of the SIR process}

\item{t_end}{end of observation period}

\item{K}{number of observation intervals}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning parameter for the Markov chain}

\item{rho}{proportion of the latent data updated each iteration}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{m}{number of independent iterations to estimate the coverage rate}

\item{a}{array number}
}
\value{
a tibble summarizing the coverage
}
\description{
Experiment 4: estimate the coverage rate of the DA-MCMC algorithm
}
