% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_4_output_analysis}
\alias{experiment_4_output_analysis}
\title{Title}
\usage{
experiment_4_output_analysis(output_E4, path)
}
\arguments{
\item{output_E4}{output from E4}

\item{path}{path for figures}
}
\value{

}
\description{
Title
}
