% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_5_ebola}
\alias{experiment_5_ebola}
\title{Experiment 5: Analysis of the Ebola data}
\usage{
experiment_5_ebola(
  I0 = 10,
  theta_0 = list(R0 = 1, gamma = 0.1),
  N = 1e+05,
  thin = 1,
  rho = 1,
  param = "bg",
  approx = "ldp",
  iota_dist = "exponential",
  gener = FALSE,
  b = 1/2,
  path = NULL,
  plot_id = "E5"
)
}
\arguments{
\item{I0}{size initial infectious population}

\item{theta_0}{intial parameter values}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning parameter for the Markov chain}

\item{rho}{proportion of the latent data updated each iteration}

\item{param}{c("bg", "bR"); parameterize the models in terms of (beta, gamma) or (beta, R0)}

\item{approx}{c("poisson", "ldp"); whether to approximate the distribution of the infection times with a poisson process or a linear death process}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}

\item{path}{directory in which to save the figure}

\item{plot_id}{name file for the figures}
}
\value{
list containing the parameters, observed data, Markov chain and run time of the algorithm
}
\description{
Experiment 5: Analysis of the Ebola data
}
