% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_5_output}
\alias{experiment_5_output}
\title{Analyze output of MCMC run from Experiment 5}
\usage{
experiment_5_output(
  output_E5,
  iota_dist = "exponential",
  path,
  plot_id = "E5",
  burnin = 0,
  thin = 1,
  n_max = NULL
)
}
\arguments{
\item{output_E5}{object returned by the function experiment_5_ebola}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{path}{directory in which to save the figure}

\item{plot_id}{name file for the figure(s)}

\item{burnin}{number of iterations to discard, the default value NULL will discard half of the draws.}

\item{thin}{thinning parameter for the Markov chain}

\item{n_max}{last draw to take into account}
}
\value{
figures and list of summary statistics
}
\description{
Analyze output of MCMC run from Experiment 5
}
