% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_I.R
\name{extract_Ebola_data}
\alias{extract_Ebola_data}
\title{Extract Observed Infection Counts from .csv File}
\usage{
extract_Ebola_data(path)
}
\arguments{
\item{path}{directory in which the .csv file is located}
}
\value{
Infection counts for Gueckedou
}
\description{
Extract Observed Infection Counts from .csv File
}
