% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_parameters.R
\name{gibbs_beta_gamma}
\alias{gibbs_beta_gamma}
\title{Gibbs sampler for beta and gamma}
\usage{
gibbs_beta_gamma(SS, par_prior, Y)
}
\arguments{
\item{SS}{sufficient statistics of some latent data}

\item{par_prior}{parameters of the prior distribution}

\item{Y}{observed data}
}
\value{
draws from the joint full conditional for beta and gamma
}
\description{
Gibbs sampler for beta and gamma
}
