% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_beta_R0.R
\name{gibbs_beta_R0}
\alias{gibbs_beta_R0}
\title{Two-stage Gibbs sampler for beta and R0 for the stochastic SIR model}
\usage{
gibbs_beta_R0(SS, par_prior, theta, Y)
}
\arguments{
\item{SS}{sufficient statistics#'}

\item{par_prior}{parameters of the prior distribution}

\item{theta}{parameters of the SIR process}

\item{Y}{observed data}
}
\value{
new value for (beta, R0)
}
\description{
Two-stage Gibbs sampler for beta and R0 for the stochastic SIR model
}
