% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_theta.R
\name{gibbs_theta}
\alias{gibbs_theta}
\title{Gibbs sampler for theta}
\usage{
gibbs_theta(SS, iota_dist, par_prior, theta, param, Y)
}
\arguments{
\item{SS}{sufficient statistics}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{par_prior}{parameters of the prior distributions}

\item{theta}{parameters of the SIR process}

\item{param}{c("bg", "bR"); parameterization (beta, gamma) or (beta, R0)}

\item{Y}{observed data}
}
\value{
new value for theta
}
\description{
allows two parameterizations of the stochastic SIR model
}
