% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_latent.R
\name{propose_tau_I}
\alias{propose_tau_I}
\title{Generates infection times}
\usage{
propose_tau_I(n, mu, lower, upper, approx)
}
\arguments{
\item{n}{number of infection times to generate}

\item{mu}{individual rate of infection}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{approx}{poisson approximation}
}
\value{
infection times
}
\description{
Generates infection times
}
