% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_tau.R
\name{propose_tau_J}
\alias{propose_tau_J}
\title{Generates removal times}
\usage{
propose_tau_J(theta, tau_T, iota_dist, t_end)
}
\arguments{
\item{theta}{current values of the parameters for the SIR}

\item{tau_T}{observed event times}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{t_end}{time before which individuals may be removed}
}
\value{
removal times
}
\description{
Generates removal times
}
