% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_latent.R
\name{propose_tau_R}
\alias{propose_tau_R}
\title{Generates removal times}
\usage{
propose_tau_R(theta, tau_I, iota_dist, t_end)
}
\arguments{
\item{theta}{parameters}

\item{tau_I}{observed event times}

\item{iota_dist}{distribution of infection periods}

\item{t_end}{time before which individuals may be removed}
}
\value{
removal times
}
\description{
Generates removal times
}
