% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_weibull2.R
\name{pweibull2}
\alias{pweibull2}
\title{CDF of weibull distribution}
\usage{
pweibull2(x, shape, lambda, log.p = FALSE, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{shape}{shape parameter}

\item{lambda}{rate parameter}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x), otherwise, P(X > x)}
}
\value{
vector of tail probabilities
}
\description{
Lower and upper tail probabilities of the weibull distribution with the shape and rate parameterization.
}
\seealso{
Other weibull2: 
\code{\link{dweibull2_trunc_log}()},
\code{\link{dweibull2}()},
\code{\link{rweibull2_trunc}()},
\code{\link{rweibull2}()}
}
\concept{weibull2}
