% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_to_scale.R
\name{rate_to_scale}
\alias{rate_to_scale}
\title{Re-parameterization for the weibull distribution}
\usage{
rate_to_scale(rate, shape)
}
\arguments{
\item{rate}{rate parameter}

\item{shape}{shape parameter}
}
\value{
scale parameter
}
\description{
Provides the scale parameter of the Weibull distribution
}
