% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_burnin.R
\name{remove_burnin}
\alias{remove_burnin}
\title{Removes the iterations of the transient phase of a Markov chain}
\usage{
remove_burnin(df, burnin = 0)
}
\arguments{
\item{df}{data frame}

\item{burnin}{number of iterations to remove as burnin}
}
\value{
data frame without the burnin iterations
}
\description{
Removes the iterations of the transient phase of a Markov chain
}
