% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp_trunc.R
\name{rexp_trunc}
\alias{rexp_trunc}
\title{Random generator for the truncated exponential distribution}
\usage{
rexp_trunc(n, lambda, lower, upper)
}
\arguments{
\item{n}{integer}

\item{lambda}{positive double}

\item{lower}{positive double}

\item{upper}{positive double}
}
\value{
positive double
}
\description{
Generates random values from a truncated exponential distribution
}
\examples{
hist(
rexp_trunc(1e4, lambda = 3, lower = 1, upper = 2),
main = "Truncated exponential distribution", xlab = NULL
)

}
