% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprop_x.R
\name{rprop_x}
\alias{rprop_x}
\title{Generates PD-SIR process conditionally on the observed data}
\usage{
rprop_x(
  theta,
  Y,
  generalized,
  b,
  x = NULL,
  rho = 1,
  iota_distribution = "exponential",
  approximation = "pld"
)
}
\arguments{
\item{theta}{parameters for the SIR}

\item{Y}{observed data}

\item{generalized}{logical; whether to use the generalized SIR}

\item{b}{parameter of the generalized SIR}

\item{x}{current configuration of the latent data}

\item{rho}{proportion of the latent being updated}

\item{iota_distribution}{c("exponential", "weibull"); distribution of the infection period}

\item{approximation}{c("poisson", "lbd"); whether to approximate the distribution of the infection times with a poisson process or a linear death process}
}
\value{
latent data for the SIR
}
\description{
Generates PD-SIR process conditionally on the observed data
}
