% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAMCMC.R
\name{run_DAMCMC}
\alias{run_DAMCMC}
\title{Run a DA-MCMC to fit the stochastic SIR model to discretely observed incidence counts with the PD-SIR algorithm}
\usage{
run_DAMCMC(
  Y,
  rho = 1/10,
  N = 10000,
  thin = 1,
  theta_0,
  par_prior = list(a_beta = 0.1, b_beta = 1, a_gamma = 1, b_gamma = 1, a_R0 = 2, b_R0 =
    0.002),
  parameterization = "bR",
  generalized = FALSE,
  b = 1/2,
  print_i = FALSE,
  save_SS = FALSE
)
}
\arguments{
\item{Y}{observed data}

\item{rho}{portion of the latent being updated each iteration}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning argument for the Markov chain}

\item{theta_0}{initial value for the parameters}

\item{par_prior}{parameters of the prior distribution}

\item{parameterization}{c("bg", "bR"); parameterize the models in terms of (beta, gamma) or (beta, R0)}

\item{generalized}{logical; whether to use the generalized SIR}

\item{b}{parameter of the generalized SIR}

\item{print_i}{logical; whether to print the iteration}

\item{save_SS}{logical; whether to save the sufficient statistics generated each iteration}
}
\value{
list with the draws for the parameters, the log likelihood, acceptance rate and size of initial susceptible population
}
\description{
Run a DA-MCMC to fit the stochastic SIR model to discretely observed incidence counts with the PD-SIR algorithm
}
