% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_weibull2.R
\name{rweibull2}
\alias{rweibull2}
\title{Random generator for the truncated weibull distribution}
\usage{
rweibull2(n, shape, rate)
}
\arguments{
\item{n}{number of random values to generate}

\item{shape}{shape parameter}

\item{rate}{rate parameter}
}
\value{
vector of n random variables following a truncated weibull distribution
}
\description{
shape and rate parameterization.
}
\seealso{
Other weibull2: 
\code{\link{dweibull2_trunc_log}()},
\code{\link{dweibull2}()},
\code{\link{pweibull2}()},
\code{\link{rweibull2_trunc}()}
}
\concept{weibull2}
