% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{rweibull2_trunc}
\alias{rweibull2_trunc}
\title{Random generator for the truncated weibull distribution}
\usage{
rweibull2_trunc(n, shape, rate, lower, upper)
}
\arguments{
\item{n}{number of random variables to generate.}

\item{shape}{shape parameter}

\item{rate}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
n-length vector of random variables following a truncated weibull distribution
}
\description{
Random generator for the truncated weibull distribution
}
\seealso{
Other truncated weibull: 
\code{\link{dweibull2_trunc_log}()}
}
\concept{truncated weibull}
