% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_iota.R
\name{simulate_iota}
\alias{simulate_iota}
\title{Simulate infection periods}
\usage{
simulate_iota(n, iota_dist, gamma, shape, rate)
}
\arguments{
\item{n}{number of infection periods to simulate}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gamma}{parameter of the exponential distribution}

\item{shape}{parameter of the weibull distribution}

\item{rate}{parameter of the weibull distribution}
}
\value{
vector of random infectious durations
}
\description{
Simulate infection periods
}
