% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SIR.R
\name{simulate_SIR}
\alias{simulate_SIR}
\title{Simulate trajectories of the Stochastic SIR and SEIR processes}
\usage{
simulate_SIR(
  S0 = 1000,
  I0 = 10,
  theta = list(R0 = 2, gamma = 1),
  t_end = 6,
  generalized = FALSE,
  b = 1/2,
  type = "SIR"
)
}
\arguments{
\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{theta}{parameters of the SIR process}

\item{t_end}{end of observation period}

\item{generalized}{logical; whether to use generalized SIR}

\item{b}{parameter of generalized SIR}

\item{type}{c("SIR", "SEIR"); type of process to simulate}
}
\value{
list with all types of useful objects
}
\description{
Also computes additional statistics such as MLEs and summary statistics
}
