% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficient_statistics.R
\name{sufficient_statistics}
\alias{sufficient_statistics}
\title{Compute the sufficient statistics from the latent space}
\usage{
sufficient_statistics(x, Y, generalized, b, return_SI = FALSE)
}
\arguments{
\item{x}{latent data}

\item{Y}{observed data}

\item{generalized}{logical; whether to use the generalized SIR}

\item{b}{parameter of the generalized SIR}

\item{return_SI}{logical; whether to return the trajectories of S and I}
}
\value{
sufficient statistics
}
\description{
These sufficient statistics are used to conduct inference on the parameters
}
