% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_MCMC.R
\name{analyze_MCMC}
\alias{analyze_MCMC}
\title{Creates figures and summary statistics of the output of the Markov chain}
\usage{
analyze_MCMC(
  MC,
  burnin = 0,
  thin = 1,
  plot_id = NULL,
  path = NULL,
  save_fig = TRUE,
  do_SS = FALSE,
  beta_true,
  gamma_true
)
}
\arguments{
\item{MC}{output of the MCMC algorithm}

\item{burnin}{number of iterations to discard}

\item{thin}{thinning argument for the iterations of the Markov chain}

\item{plot_id}{name file for the figures}

\item{path}{directory in which to save the figure}

\item{save_fig}{logical; whether to save the figures generated}

\item{do_SS}{logical; whether to analyze the summary statistics in addition to the parameters}

\item{beta_true}{true value of the beta parameter}

\item{gamma_true}{true value of the gamma parameter}
}
\value{
a list of summary statistics including posterior means, posterior quantiles, effective sample size and the acceptance rate
}
\description{
Creates figures and summary statistics of the output of the Markov chain
}
