% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pi.R
\name{compute_pi}
\alias{compute_pi}
\title{Probability of being removed before next observation time}
\usage{
compute_pi(theta, tau_T, iota_distribution, t_end)
}
\arguments{
\item{theta}{named list with the parameters gamma (and nu and lambda)}

\item{tau_T}{observed event times}

\item{iota_distribution}{c("exponential", "weibull")}

\item{t_end}{time before which individuals may be removed}
}
\value{
a vector of the probabilities that individual infected at time tau_T are removed before t_end
}
\description{
Compute p_i, the probability of being removed before t_end given an infection at time tau_T.
}
