% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Tk.R
\name{compute_Tk}
\alias{compute_Tk}
\title{Discrete Incidence Data for Infections}
\usage{
compute_Tk(x, ts, S0)
}
\arguments{
\item{x}{vector of infection times}

\item{ts}{observation schedule}

\item{S0}{initial population of susceptibles}
}
\value{
a vector of the number of infections in each time interval
}
\description{
Compute the number of infections in each time interval from the output of a MCMC
}
