% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution_observed_removal}
\alias{contribution_observed_removal}
\title{Log proposal density of observed removals}
\usage{
contribution_observed_removal(
  theta,
  tau_T,
  tau_J,
  tau_J_obs,
  iota_distribution,
  t_end
)
}
\arguments{
\item{theta}{parameters of SIR model}

\item{tau_T}{infection times}

\item{tau_J}{removal times}

\item{tau_J_obs}{index of observed removal times}

\item{iota_distribution}{c("exponential", "weibull"); distribution of infectious time}

\item{t_end}{end of observation window}
}
\value{
log density
}
\description{
Log proposal density of observed removals
}
