% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_latent.R
\name{dprop_x}
\alias{dprop_x}
\title{Log proposal density for the latent data}
\usage{
dprop_x(theta, Y, x, i_update, gener, b, iota_dist, approx)
}
\arguments{
\item{theta}{current values of the parameters for the SIR}

\item{Y}{observed data}

\item{x}{current configuration of the latent data}

\item{i_update}{index set of particles whose values are updated}

\item{gener}{generalized}

\item{b}{parameter of generalized}

\item{iota_dist}{distribution of infection periods}

\item{approx}{poisson approximation}
}
\value{
log density
}
\description{
Log proposal density for the latent data
}
