% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprop_x.R
\name{dprop_x}
\alias{dprop_x}
\title{Log proposal density for the latent data}
\usage{
dprop_x(
  theta,
  Y,
  x,
  i_update,
  generalized,
  b,
  iota_distribution = "exponential",
  approximation = "ldp"
)
}
\arguments{
\item{theta}{parameters of the SIR process}

\item{Y}{observed data}

\item{x}{latent data}

\item{i_update}{index set of particles whose values are updated}

\item{generalized}{logical; whether to consider the generalized SIR process}

\item{b}{parameter of generalized SIR process}

\item{iota_distribution}{c("exponential", "weibull"); distribution of the infection period}

\item{approximation}{c("poisson", "pld"); type of proposal process}
}
\value{
log density
}
\description{
Log proposal density for the latent data
}
