% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM-process.R
\name{draw_trajectories}
\alias{draw_trajectories}
\title{Draw trajectories of a SEM}
\usage{
draw_trajectories(
  SEM,
  t_end = 10,
  xlab = "Time",
  text_size = 15,
  include_legend = TRUE
)
}
\arguments{
\item{SEM}{list corresponding to stochastic epidemic process}

\item{t_end}{end time of the simulation period}

\item{xlab}{label of x-axis}

\item{text_size}{size of text in figure}

\item{include_legend}{logical; wehther to include a legend}
}
\value{
save a figure of the compartments' trajectories
}
\description{
Line plot of the compartment sizes over time.
}
