% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SEM.R
\name{draw_trajectories}
\alias{draw_trajectories}
\title{Draw trajectories of a SEM}
\usage{
draw_trajectories(
  SEM,
  t_end = 10,
  save_figure = TRUE,
  path_figure = NULL,
  id_figure = NULL,
  xlab = "time"
)
}
\arguments{
\item{SEM}{list corresponding to stochastic epidemic process}

\item{t_end}{end time of the simulation period}

\item{save_figure}{logical; whether to save the figure}

\item{path_figure}{path to save figure}

\item{id_figure}{logical; whether to save the figure}

\item{xlab}{label of x-axis}
}
\value{
save a figure of the compartments' trajectories
}
\description{
Line plot of the compartment sizes over time.
}
