% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_weibull2.R
\name{dweibull2}
\alias{dweibull2}
\title{Density of weibull distribution}
\usage{
dweibull2(x, shape, lambda, log = FALSE)
}
\arguments{
\item{x}{vector of values}

\item{shape}{shape parameter}

\item{lambda}{rate parameter}

\item{log}{logical; whether to take the logarithm of the density}
}
\value{
vector of densities
}
\description{
Density of weibull distribution
}
\seealso{
Other weibull2: 
\code{\link{dweibull2_trunc_log}()},
\code{\link{pweibull2}()},
\code{\link{rweibull2_trunc}()},
\code{\link{rweibull2}()}
}
\concept{weibull2}
