% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{dweibull2_trunc_log}
\alias{dweibull2_trunc_log}
\title{Log density of the truncated Weibull distribution}
\usage{
dweibull2_trunc_log(x, shape, rate, lower, upper)
}
\arguments{
\item{x}{vector}

\item{shape}{shape parameter}

\item{rate}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
vector of densities
}
\description{
Log density of the truncated Weibull distribution
}
\seealso{
Other truncated weibull: 
\code{\link{rweibull2_trunc}()}
}
\concept{truncated weibull}
