% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_weibull2.R
\name{dweibull2_trunc_log}
\alias{dweibull2_trunc_log}
\title{Log density of the truncated Weibull distribution}
\usage{
dweibull2_trunc_log(x, shape, rate, lower, upper)
}
\arguments{
\item{x}{vector of values}

\item{shape}{shape parameter}

\item{rate}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
vector of log densities
}
\description{
Log density of the truncated Weibull distribution
}
\seealso{
Other weibull2: 
\code{\link{dweibull2}()},
\code{\link{pweibull2}()},
\code{\link{rweibull2_trunc}()},
\code{\link{rweibull2}()}
}
\concept{weibull2}
