% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_1_output_analysis}
\alias{experiment_1_output_analysis}
\title{Analyze output of MCMC run from Experiment 1}
\usage{
experiment_1_output_analysis(
  x,
  iota_dist = "exponential",
  theta_true,
  burnin = NULL,
  n_max = NULL,
  plot_id = NULL,
  path = NULL,
  save_fig = TRUE
)
}
\arguments{
\item{x}{object returned by the function experiment_1_proof_of_concept}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{theta_true}{true value of the parameters}

\item{burnin}{number of iterations to discard, the default value NULL will discard half of the draws.}

\item{n_max}{last draw to take into account}

\item{plot_id}{name file for the figure(s)}

\item{path}{directory in which to save the figure}

\item{save_fig}{logical; whether to save the figures generated}
}
\value{
figures and list of summary statistics with and without burn-in
}
\description{
Analyze output of MCMC run from Experiment 1
}
