% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_1_output_analysis}
\alias{experiment_1_output_analysis}
\title{Analyze output of MCMC run}
\usage{
experiment_1_output_analysis(x, plot_id = NULL, path = NULL, burnin = NULL)
}
\arguments{
\item{x}{object returned by the function experiment_1_proof_of_concept}

\item{plot_id}{name file for the figures}

\item{path}{directory in which to save the figure}

\item{burnin}{number of iterations to discard, the default value NULL will discard half of the draws.}
}
\value{
list of summary statistics with and without burnin
}
\description{
Analyze output of MCMC run
}
