% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_1_proof_of_concept}
\alias{experiment_1_proof_of_concept}
\title{Experiment 1: Proof of concept}
\usage{
experiment_1_proof_of_concept(
  S0 = 1000,
  I0 = 10,
  theta = list(R0 = 2.5, gamma = 1),
  t_end = 6,
  K = 10,
  N = 10000,
  thin = 1,
  rho = 1/5,
  theta_0_factor_beta = 0.1,
  theta_0_factor_gamma = 0.1
)
}
\arguments{
\item{S0}{size of initial susceptible population}

\item{I0}{size initial infectious population}

\item{theta}{parameters of the SIR process}

\item{t_end}{end of observation period}

\item{K}{number of observation intervals}

\item{N}{number of iterations of the Markov chain}

\item{thin}{thinning parameter for the Markov chain}

\item{rho}{proportion of the latent data updated each iteration}

\item{theta_0_factor_beta}{factor by which the true value of beta is multiplied to determine the starting point of the Markov chain}

\item{theta_0_factor_gamma}{factor by which the true value of beta is multiplied to determine the starting point of the Markov chain}
}
\value{
list containing the parameters, observed data, Markov chain and run time of the algorithm
}
\description{
Experiment 1: Proof of concept
}
