% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_log.R
\name{f_log}
\alias{f_log}
\title{Log likelihood of the stochastic SIR model}
\usage{
f_log(theta, SS, generalized, b, iota_distribution = "exponential")
}
\arguments{
\item{theta}{parameters of the stochastic SIR process}

\item{SS}{sufficient statistics of some latent data}

\item{generalized}{logical; whether to use the generalized stochastic SIR process (Severo, 1969)}

\item{b}{parameter of the generalized SIR}

\item{iota_distribution}{c("exponential", "weibull"); distribution of the infection period}
}
\value{
log likelihood
}
\description{
Log likelihood of the stochastic SIR model
}
