% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_theta.R
\name{gibbs_theta}
\alias{gibbs_theta}
\title{Gibbs sampler for theta}
\usage{
gibbs_theta(SS, par_prior, theta, parameterization, Y)
}
\arguments{
\item{SS}{sufficient statistics}

\item{par_prior}{parameters of the prior distribution}

\item{theta}{parameters of the SIR process}

\item{parameterization}{c("bg", "bR"); parameterize the models in terms of (beta, gamma) or (beta, R0)}

\item{Y}{observed data}
}
\value{
new value for theta
}
\description{
allows two parameterizations of the stochastic SIR model
}
