% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_tau_J.R
\name{propose_tau_J}
\alias{propose_tau_J}
\title{Generates removal times}
\usage{
propose_tau_J(theta, tau_T, iota_distribution, t_end)
}
\arguments{
\item{theta}{parameters of the SIR process}

\item{tau_T}{infection times}

\item{iota_distribution}{c("exponential", "weibull"); distribution of infection period}

\item{t_end}{end of observation period}
}
\value{
removal times
}
\description{
Generates removal times
}
