% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{pweibull2}
\alias{pweibull2}
\title{CDF of weibull distribution}
\usage{
pweibull2(x, shape, rate, log.p = FALSE, lower.tail = TRUE)
}
\arguments{
\item{x}{vector}

\item{shape}{shape parameter}

\item{rate}{rate parameter}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x), otherwise, P(X > x)}
}
\value{
vector of tail probabilities
}
\description{
CDF of weibull distribution
}
