% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_exp.R
\name{rexp_trunc}
\alias{rexp_trunc}
\title{Random generator for the truncated exponential distribution}
\usage{
rexp_trunc(n, lambda, lower, upper)
}
\arguments{
\item{n}{number of random values to generates}

\item{lambda}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
positive double
}
\description{
Generates random values from a truncated exponential distribution
}
\examples{
hist(
rexp_trunc(1e4, lambda = 3, lower = 1, upper = 2),
main = "Truncated exponential distribution", xlab = NULL
)

}
\seealso{
Other truncated exponential: 
\code{\link{dexp_trunc_log}()}
}
\concept{truncated exponential}
