% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_latent.R
\name{rprop_x}
\alias{rprop_x}
\title{Generates PD-SIR process conditionally on the observed data}
\usage{
rprop_x(
  theta,
  Y,
  gener = FALSE,
  b = 1,
  iota_dist,
  approx = "ldp",
  x = NULL,
  rho = 1
)
}
\arguments{
\item{theta}{current values of the parameters for the SIR}

\item{Y}{observed data}

\item{gener}{generalized}

\item{b}{parameter of generalized}

\item{iota_dist}{distribution of infection periods}

\item{approx}{poisson approximation}

\item{x}{current configuration of the latent data}

\item{rho}{proportion}
}
\value{
latent data for the SIR
}
\description{
Generates PD-SIR process conditionally on the observed data
}
