% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAMCMC.R
\name{run_DAMCMC}
\alias{run_DAMCMC}
\title{Run a DA-MCMC to fit the stochastic SIR model to discretely observed incidence counts with the PD-SIR algorithm}
\usage{
run_DAMCMC(
  Y,
  N = 10000,
  rho = 1,
  param = "bg",
  approx = "ldp",
  iota_dist = "exponential",
  gener = FALSE,
  b = 1/2,
  thin = 1,
  theta_0,
  print_i = FALSE,
  save_SS = FALSE,
  par_prior = list(a_beta = 0.01, b_beta = 1, a_gamma = 1, b_gamma = 1, a_R0 = 2, b_R0
    = 2, a_lambda = 1, b_lambda = 1)
)
}
\arguments{
\item{Y}{observed data}

\item{N}{number of iterations of the Markov chain}

\item{rho}{proportion of the latent data updated each iteration}

\item{param}{c("bg", "bR"); parameterize the models in terms of (beta, gamma) or (beta, R0)}

\item{approx}{c("poisson", "ldp"); whether to approximate the distribution of the infection times with a poisson process or a linear death process}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}

\item{thin}{thinning parameter for the Markov chain}

\item{theta_0}{initial value for the parameters}

\item{print_i}{logical; whether to print the iteration}

\item{save_SS}{logical; whether to save the sufficient statistics generated each iteration}

\item{par_prior}{parameters of the prior distributions}
}
\value{
list with the draws for the parameters, the log likelihood, acceptance rate and size of initial susceptible population
}
\description{
Run a DA-MCMC to fit the stochastic SIR model to discretely observed incidence counts with the PD-SIR algorithm
}
