% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_weibull2.R
\name{rweibull2_trunc}
\alias{rweibull2_trunc}
\title{Random generator for the truncated weibull distribution}
\usage{
rweibull2_trunc(n, shape, lambda, lower, upper)
}
\arguments{
\item{n}{number of infection times to generate}

\item{shape}{shape parameter}

\item{lambda}{rate parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
vector of n random variables following a truncated weibull distribution
}
\description{
Random generator for the truncated weibull distribution
}
\seealso{
Other weibull2: 
\code{\link{dweibull2_trunc_log}()},
\code{\link{dweibull2}()},
\code{\link{pweibull2}()},
\code{\link{rweibull2}()}
}
\concept{weibull2}
