% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_iota}
\alias{simulate_iota}
\title{Simulate infection periods}
\usage{
simulate_iota(n, iota_dist, gamma, shape, lambda)
}
\arguments{
\item{n}{number of infection periods to simulate}

\item{iota_dist}{distribution of infection periods}

\item{gamma}{parameter of the exponential distribution}

\item{shape}{parameter of the weibull distribution}

\item{lambda}{parameter of the weibull distribution}
}
\value{
vector of random infectious durations
}
\description{
Simulate infection periods
}
