% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM-simulate.R
\name{simulate_SEM}
\alias{simulate_SEM}
\title{Simulate trajectories of the Stochastic SIR and SEIR processes}
\usage{
simulate_SEM(
  S0 = 1000,
  I0 = 10,
  t_end = 6,
  theta = list(R0 = 2.5, gamma = 1, lambda = 1, shape = 1),
  iota_dist = "exponential",
  gener = FALSE,
  b = 1/2
)
}
\arguments{
\item{S0}{intial exposed population size}

\item{I0}{intial exposed population size}

\item{t_end}{end time of the simulation period}

\item{theta}{parameters}

\item{iota_dist}{c("exponential", "weibull"); distribution of the infection period}

\item{gener}{logical; whether to use the generalized SIR of Severo (1972)}

\item{b}{parameter of the generalized SIR}
}
\value{
a list with all types of useful objects
}
\description{
Also computes additional statistics such as MLEs and summary statistics
}
