% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficient-statistics.R
\name{suff_stat}
\alias{suff_stat}
\title{Compute the sufficient statistics from the latent space}
\usage{
suff_stat(x, Y, gener, b, return_SI = FALSE)
}
\arguments{
\item{x}{current configuration of the latent data}

\item{Y}{observed data}

\item{gener}{generalized}

\item{b}{parameter of generalized}

\item{return_SI}{logical; whether to return the trajectories of S and I}
}
\value{
sufficient statistics
}
\description{
These sufficient statistics are used to conduct inference on the parameters
}
